Move and rename search function modules
