--- src/search-file/index.js
@@ -1,3 +1,0 @@
-import searchFile from './search-file'
-
-export default searchFile
