--- src/search-older-file-types/index.js
@@ -1,3 +1,0 @@
-import searchOlderFileTypes from './search-older-file-types'
-
-export default searchOlderFileTypes
