+++ src/text-search/index.js
@@ -1,0 +1,6 @@
+import textSearch from './text-search'
+import searchOlderFileTypes from './search-older-file-types'
+
+textSearch.older = searchOlderFileTypes
+
+export default textSearch
