+++ src/text-search/search-older-file-types.js
