+++ src/text-search/text-search.js
