+++ src/text-search/text-search.test.js
@@ -1,12 +1,12 @@
 import test from 'ava'
-import searchFile from './search-file'
+import textSearch from './text-search'
 
 test('Able to read .docx format', async t => {
   const file = 'test_source/the-great-sea.docx'
   const searchTerms = ['then', 'this']
   let found = null
   try {
-    found = await searchFile({file, searchTerms})
+    found = await textSearch({file, searchTerms})
   }
   catch (error) {
     t.fail(error)
